+++ R/monitor_join.R
@@ -7,7 +7,7 @@
 #' @return A \emph{ws_monitor} object with merged timeseries.
 #' @description For each monitor in \code{monitorIDs}, an attempt is made to merge
 #' the associated data from \code{ws_monitor1} and \code{ws_monitor2} and.
-#' 
+#'
 #' This is useful when the same \code{monitorID} appears in different \emph{ws_monitor}
 #' objects representing different time periods. The returned \emph{ws_monitor}
 #' object will cover both time periods.
@@ -21,10 +21,10 @@
 #'                       timezone='America/Los_Angeles')
 #' Methow_Valley <- monitor_join(Jul, Aug, monitorIDs=c('530470010_01','530470009_01'))
 #' }
-monitor_join <- function(ws_monitor1=NULL,
-                         ws_monitor2=NULL,
-                         monitorIDs=NULL) {
-  
+monitor_join <- function(ws_monitor1 = NULL,
+                         ws_monitor2 = NULL,
+                         monitorIDs = NULL) {
+
   # Sanity checks
   if ( !"ws_monitor" %in% class(ws_monitor1) ) stop("Required argument 'ws_monitor1' is not a ws_monitor object")
   if ( !"ws_monitor" %in% class(ws_monitor2) ) stop("Required argument 'ws_monitor2' is not a ws_monitor object")
@@ -32,15 +32,15 @@ monitor_join <- function(ws_monitor1=NULL,
     monitorIDs <- intersect(ws_monitor1$meta$monitorID, ws_monitor2$meta$monitorID)
   }
 
-  # Create an overall time axis  
+  # Create an overall time axis
   starttime <- min(ws_monitor1$data$datetime, ws_monitor2$data$datetime)
   endtime <- max(ws_monitor1$data$datetime, ws_monitor2$data$datetime)
   datetime <- seq(starttime, endtime, by="hours")
   hourlyDF <- data.frame(datetime=datetime)
-  
+
   monList <- list()
   for ( monitorID in monitorIDs ) {
-    
+
     # Get trimmed, single version of mon1
     mon1 <- monitor_subset(ws_monitor1, monitorIDs=monitorID)
     if ( monitor_isEmpty(mon1) ) {
@@ -51,7 +51,7 @@ monitor_join <- function(ws_monitor1=NULL,
       data1 <- dplyr::full_join(hourlyDF, mon1$data, by="datetime")
     }
     mask1 <- !is.na(data1[[monitorID]])
-    
+
     # Get trimmed, single version of mon2
     mon2 <- monitor_subset(ws_monitor2, monitorIDs=monitorID)
     if ( monitor_isEmpty(mon2) ) {
@@ -62,9 +62,9 @@ monitor_join <- function(ws_monitor1=NULL,
       data2 <- dplyr::full_join(hourlyDF, mon2$data, by="datetime")
     }
     mask2 <- !is.na(data2[[monitorID]])
-    
+
     mask3 <- mask1 & mask2
-    
+
     # Join the data
     joinedData <- data1
     joinedData[[monitorID]][mask2] <- data2[[monitorID]][mask2]
@@ -76,22 +76,21 @@ monitor_join <- function(ws_monitor1=NULL,
       }
     }
 
-    # Get metadata (while handling the possibility that one of the monitor objects might be empty)
-    if ( nrow(mon1$meta) == 1 ) {
+    # Create a valid monitor object if possible
+    if ( !monitor_isEmpty(mon1) ) {
       mon <- list(meta=mon1$meta, data=joinedData)
-    } else if ( nrow(mon2$meta == 1) ) {
+      monList[[monitorID]] <- structure(mon, class = c("ws_monitor", "list"))
+    } else if ( !monitor_isEmpty(mon2) ) {
       mon <- list(meta=mon2$meta, data=joinedData)
+      monList[[monitorID]] <- structure(mon, class = c("ws_monitor", "list"))
     } else {
-      stop(paste0("monitor '", monitorID,"' is not found in ws_monitor1 or ws_monitor2"))
+      # Bad monitor ID. Do nothing.
     }
-    
-    # Create valid monitor object
-    monList[[monitorID]] <- structure(mon, class = c("ws_monitor", "list"))
-    
-  }  
-  
+
+  }
+
   ws_monitor <- monitor_combine(monList)
-  
+
   return(ws_monitor)
-  
+
 }